+++ Source/SDK/Api/OAuthTokenCredential.cs
@@ -78,6 +78,14 @@ namespace PayPal.Api
         /// then a new token will be created when calling <see cref="OAuthTokenCredential.GetAccessToken()"/>.
         /// </summary>
         public int AccessTokenExpirationSafetyGapInSeconds { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="config"></param>
+        public OAuthTokenCredential(Dictionary<string, string> config) : this(string.Empty, string.Empty, config)
+        {
+        }
                
         /// <summary>
         /// Client Id and Secret for the OAuth
@@ -95,9 +103,11 @@ namespace PayPal.Api
         /// <param name="clientSecret"></param>
         public OAuthTokenCredential(string clientId, string clientSecret, Dictionary<string, string> config)
         {
-            this.ClientId = clientId;
-            this.ClientSecret = clientSecret;
-            this.config = config != null ? ConfigManager.GetConfigWithDefaults(config) : ConfigManager.GetConfigWithDefaults(ConfigManager.Instance.GetProperties()); 
+            this.config = config != null ? ConfigManager.GetConfigWithDefaults(config) : ConfigManager.GetConfigWithDefaults(ConfigManager.Instance.GetProperties());
+
+            this.ClientId = string.IsNullOrEmpty(clientId) ? (this.config.ContainsKey(BaseConstants.ClientId) ? this.config[BaseConstants.ClientId] : string.Empty) : clientId;
+            this.ClientSecret = string.IsNullOrEmpty(clientSecret) ? (this.config.ContainsKey(BaseConstants.ClientSecret) ? this.config[BaseConstants.ClientSecret] : string.Empty) : clientSecret;
+
             this.SdkVersion = new SDKVersion();
             this.AccessTokenExpirationSafetyGapInSeconds = 120; // Default is 2 minute safety gap for token expiration.
         }
@@ -138,6 +148,12 @@ namespace PayPal.Api
                 // computing the token
                 // Set the Value inside the accessToken and result
                 var base64ClientId = OAuthTokenCredential.ConvertClientCredentialsToBase64String(this.ClientId, this.ClientSecret);
+
+                logger.DebugFormat("\n" +
+                    "  ClientID:       {0}\n" +
+                    "  ClientSecret:   {1}\n" +
+                    "  Base64ClientId: {2}", this.ClientId, this.ClientSecret, base64ClientId);
+
                 this.accessToken = this.GenerateOAuthToken(base64ClientId);
             }
             return this.accessToken;
@@ -228,6 +244,8 @@ namespace PayPal.Api
             httpRequest.Method = "POST";
             httpRequest.Accept = "*/*";
             httpRequest.ContentType = "application/x-www-form-urlencoded";
+
+            // Set User-Agent HTTP header
             var userAgentMap = UserAgentHeader.GetHeader();
             foreach (KeyValuePair<string, string> entry in userAgentMap)
             {
@@ -237,9 +255,16 @@ namespace PayPal.Api
                 var bytes = Encoding.Convert(Encoding.UTF8,iso8851, Encoding.UTF8.GetBytes(entry.Value));                
                 httpRequest.UserAgent = iso8851.GetString(bytes);
             }
+
+            // Set Custom HTTP headers
             foreach (KeyValuePair<string, string> header in headers)
             {
                 httpRequest.Headers.Add(header.Key, header.Value);
+            }
+
+            foreach (string headerName in httpRequest.Headers)
+            {
+                logger.DebugFormat(headerName + ":" + httpRequest.Headers[headerName]);
             }
 
             HttpConnection httpConnection = new HttpConnection(config);