+++ Source/SDK/Properties/AssemblyInfo.cs
@@ -14,9 +14,9 @@ using System.Security;
 [assembly: AssemblyTitle("PayPal .NET SDK")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("eBay Inc.")]
+[assembly: AssemblyCompany("PayPal")]
 [assembly: AssemblyProduct("PayPal")]
-[assembly: AssemblyCopyright("Copyright © 2009-2014 eBay Inc.")]
+[assembly: AssemblyCopyright("Copyright © 2009-2014 PayPal")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -38,5 +38,5 @@ using System.Security;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+[assembly: AssemblyVersion("1.1.0.0")]
+[assembly: AssemblyFileVersion("1.1.0.0")]